refactor platform
