--- lib/src/utils/platform/distro/mod.rs
@@ -1,6 +1,0 @@
-pub mod check_distro;
-pub use check_distro::check_distro;
-pub mod check_distro_result;
-pub use check_distro_result::check_distro_result;
-pub mod install_distro_packages;
-pub use install_distro_packages::install_distro_packages;
