+++ lib/src/utils/platform/mod.rs
@@ -1,6 +1,6 @@
 pub mod distro;
 pub use distro::*;
+pub mod install;
+pub use install::*;
 pub mod packages;
 pub use packages::*;
-pub mod check_platform;
-pub use check_platform::check_platform;
