--- lib/src/utils/platform/packages/install/mod.rs
@@ -1,8 +1,0 @@
-pub mod apt_install;
-pub use apt_install::apt_install;
-pub mod yum_install;
-pub use yum_install::yum_install;
-pub mod install_packages;
-pub use install_packages::install_packages;
-pub mod install_package;
-pub use install_package::install_package;
