--- lib/src/utils/platform/packages/mod.rs
@@ -1,8 +1,0 @@
-pub mod install;
-pub use install::*;
-pub mod check_package;
-pub use check_package::check_package;
-pub mod setup_packages;
-pub use setup_packages::setup_packages;
-pub mod update;
-pub use update::update;
